[monodroid] Refactor: OSBridge placed in separate files
The `OSBridge` class now lives in separate source files. No functional
changes made.
