+++ src/monodroid/jni/globals.cc
@@ -6,6 +6,7 @@ using namespace xamarin::android::internal;
 DylibMono monoFunctions;
 Util utils;
 AndroidSystem androidSystem;
+OSBridge osBridge;
 
 #ifdef DEBUG
 Debug debug;
